not truncate the score

+++ lib/uk/ac/ebi/interpro/ProcessOutputXML.groovy
@@ -471,7 +471,7 @@ class ProcessOutputXML {
             locationAttributes.each { k, v ->
                 if (v != null) {
                     if (v instanceof Number) {
-                        gen.writeNumberField(k.toString(), ((Number) v).longValue())
+                        gen.writeNumberField(k.toString(), ((Number) v).doubleValue())
                     } else {
                         gen.writeStringField(k.toString(), v.toString())
                     }
